Use double buffering and convert image locally
Use double buffering on the JavaFX Application Thread to avoid modifying
an image in use by the QuantumRenderer thread; otherwise, screen tearing
is visible when there is enough competition with other threads in the
Java virtual machine.

Convert the AWT image to a JavaFX image locally in the method
ImageAnimation.handle instead of calling the more general method
ImageUtils.toFXImage.